Move extent from Buffer to Bucket
