+++ js/data/buffer.js
@@ -13,6 +13,10 @@ var assert = require('assert');
  * statically typed structs. A buffer is comprised of items. An item is comprised of a set of
  * attributes. Attributes are defined when the class is constructed.
  *
+ * Though the buffers are intended for WebGL, this class should have no formal code dependencies
+ * on WebGL. Though the buffers are populated by vector tile features, this class should have
+ * no domain knowledge about vector tiles, coordinate systems, etc.
+ *
  * @class Buffer
  * @private
  * @param options
@@ -248,22 +252,6 @@ Buffer.AttributeType = {
  * @readonly
  */
 Buffer.ELEMENT_ATTRIBUTE_TYPE = Buffer.AttributeType.UNSIGNED_SHORT;
-
-/**
- * The maximum extent of a feature that can be safely stored in the buffer.
- * In practice, all features are converted to this extent before being added.
- *
- * Positions are stored as signed 16bit integers.
- * One bit is lost for signedness to support featuers extending past the left edge of the tile.
- * One bit is lost because the line vertex buffer packs 1 bit of other data into the int.
- * One bit is lost to support features extending past the extent on the right edge of the tile.
- * This leaves us with 2^13 = 8192
- *
- * @property {number}
- * @private
- * @readonly
- */
-Buffer.EXTENT = 8192;
 
 /**
  * @property {number}
